Fixed changes during testing

+++ application.py
@@ -187,7 +187,7 @@ def render_logout():
     url=""
     result = product.get_products()
 
-    total_pages = (len(result) / 12) + 1
+    total_pages = (int)(len(result) / 12) + 1
     if(page == None or int(page) == 1):
         result = list(itertools.islice(result, 0, 12, 1))
         return render_template('product-catalog.html', product=result, len=len(result), url=url, total_pages=total_pages), 200
